update UFT container to select correct env variable (v22.07)
This change is for v22.07 of the UFT project.

The entrypoint script that is used for building the UFT image uses an
environment variable produced by sriov network device plugin. Recent
updates to sriov network device plugin has introduced a new environment
variable that is being consumed by the entrypoint script and is breaking
UFT functionality.

The script is updated to ignore the new environment variable introduced
by newer versions of device plugin (with the suffix of _INFO).